prettify

+++ contracts/interfaces/external/synthetix/ISynthetix.sol
@@ -8,46 +8,46 @@ interface ISynthetix {
     // Views
     function availableCurrencyKeys() external view returns (bytes32[] memory);
 
-    function availableSynthCount() external view returns (uint);
+    function availableSynthCount() external view returns (uint256);
 
-    function collateral(address account) external view returns (uint);
+    function collateral(address account) external view returns (uint256);
 
-    function collateralisationRatio(address issuer) external view returns (uint);
+    function collateralisationRatio(address issuer) external view returns (uint256);
 
-    function debtBalanceOf(address issuer, bytes32 currencyKey) external view returns (uint);
+    function debtBalanceOf(address issuer, bytes32 currencyKey) external view returns (uint256);
 
     function debtBalanceOfAndTotalDebt(address issuer, bytes32 currencyKey)
         external
         view
-        returns (uint debtBalance, uint totalSystemValue);
+        returns (uint256 debtBalance, uint256 totalSystemValue);
 
     function isWaitingPeriod(bytes32 currencyKey) external view returns (bool);
 
-    function maxIssuableSynths(address issuer) external view returns (uint maxIssuable);
+    function maxIssuableSynths(address issuer) external view returns (uint256 maxIssuable);
 
     function remainingIssuableSynths(address issuer)
         external
         view
         returns (
-            uint maxIssuable,
-            uint alreadyIssued,
-            uint totalSystemDebt
+            uint256 maxIssuable,
+            uint256 alreadyIssued,
+            uint256 totalSystemDebt
         );
 
     function synths(bytes32 currencyKey) external view returns (ISnxSynth);
 
     function synthsByAddress(address synthAddress) external view returns (bytes32);
 
-    function totalIssuedSynths(bytes32 currencyKey) external view returns (uint);
+    function totalIssuedSynths(bytes32 currencyKey) external view returns (uint256);
 
-    function totalIssuedSynthsExcludeEtherCollateral(bytes32 currencyKey) external view returns (uint);
+    function totalIssuedSynthsExcludeEtherCollateral(bytes32 currencyKey) external view returns (uint256);
 
-    function transferableSynthetix(address account) external view returns (uint);
+    function transferableSynthetix(address account) external view returns (uint256);
 
     // Mutative Functions
-    function burnSynths(uint amount) external;
+    function burnSynths(uint256 amount) external;
 
-    function burnSynthsOnBehalf(address burnForAddress, uint amount) external;
+    function burnSynthsOnBehalf(address burnForAddress, uint256 amount) external;
 
     function burnSynthsToTarget() external;
 
@@ -55,34 +55,34 @@ interface ISynthetix {
 
     function exchange(
         bytes32 sourceCurrencyKey,
-        uint sourceAmount,
+        uint256 sourceAmount,
         bytes32 destinationCurrencyKey
-    ) external returns (uint amountReceived);
+    ) external returns (uint256 amountReceived);
 
     function exchangeOnBehalf(
         address exchangeForAddress,
         bytes32 sourceCurrencyKey,
-        uint sourceAmount,
+        uint256 sourceAmount,
         bytes32 destinationCurrencyKey
-    ) external returns (uint amountReceived);
+    ) external returns (uint256 amountReceived);
 
     function issueMaxSynths() external;
 
     function issueMaxSynthsOnBehalf(address issueForAddress) external;
 
-    function issueSynths(uint amount) external;
+    function issueSynths(uint256 amount) external;
 
-    function issueSynthsOnBehalf(address issueForAddress, uint amount) external;
+    function issueSynthsOnBehalf(address issueForAddress, uint256 amount) external;
 
     function mint() external returns (bool);
 
     function settle(bytes32 currencyKey)
         external
         returns (
-            uint reclaimed,
-            uint refunded,
-            uint numEntries
+            uint256 reclaimed,
+            uint256 refunded,
+            uint256 numEntries
         );
 
-    function liquidateDelinquentAccount(address account, uint susdAmount) external returns (bool);
+    function liquidateDelinquentAccount(address account, uint256 susdAmount) external returns (bool);
 }
